+++ microservice-pos/src/main/java/com/cumlpeanos/pos/repository/ReciboPOSViewRepositorio.java
@@ -8,6 +8,6 @@ import java.util.Optional;
 
 public interface ReciboPOSViewRepositorio extends JpaRepository<ReciboPOSView, BigInteger> {
 
-    Optional<ReciboPOSView> findByAlmacenAndPventa(int almacen, int pventa);
+    Optional<ReciboPOSView> findByAlmacenAndPventa(Long almacen, Long pventa);
     Optional<ReciboPOSView> findByUsrLiquidaAndEmpresa(Long usrLiquida, Long empresa);
 }
