+++ microservice-pos/src/main/java/com/cumlpeanos/pos/service/IReciboPOSViewService.java
@@ -3,6 +3,6 @@ package com.cumlpeanos.pos.service;
 import com.cumlpeanos.pos.models.entity.ReciboPOSView;
 
 public interface IReciboPOSViewService {
-    ReciboPOSView findByAlmacenAndPventa(int almacen, int pventa);
+    ReciboPOSView findByAlmacenAndPventa(Long almacen, Long pventa);
     ReciboPOSView findByUsrLiquidaAndEmpresa(Long usrLiquida, Long empresa);
 }
